requirements.txt
pandas
scikit-learn
joblib
sentence-transformers
matplotlib
umap-learn
tqdm
torch
numpy
seaborn
nltk
